travaia-ai/backend/shared/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
firebase-admin==6.2.0
PyJWT==2.8.0
requests==2.31.0
tenacity==8.2.3
livekit==0.10.0
livekit-api==1.0.5
structlog==23.2.0
